pyproject.toml
[project]
name = "rag-powered-chatbot"
requires-python = ">=3.12"
version = "0.0.0"
dependencies = [
  "loguru",
  "fastapi[standard]>=0.115.12",
  "python-dotenv>=1.1.0",
  "unstructured>=0.17.2",
  "langchain>=0.3.25",
  "langchain-community>=0.3.23",
  "langchain-cohere>=0.4.4",
  "onnxruntime==1.20.1; platform_system == 'Windows'",
  "chromadb>=1.0.13",
  "langchain-chroma>=0.2.3",
  "protobuf<=3.20"
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["I", "E4", "E7", "E9", "F", "Q"]

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "session"

[dependency-groups]
dev = [
  "pre-commit",
  "ruff",
  "pytest",
  "pytest-cov",
  "pytest-asyncio>=1.0.0",
  "ruff>=0.12.9",
  "ty>=0.0.1a19"
]
